requirements.txt
# pip install -r requirements.txt
beautifulsoup4==4.11.1
matplotlib==3.6.2
numpy==1.23.4
pandas==1.5.0
Pillow==9.5.0
Requests==2.29.0
textblob==0.17.1
